requirements.txt
PyQt6==6.6.1
PyQt6_sip==13.6.0
Pillow==10.3.0
streamlit==1.33.0
